// src/roles/role.enum.ts
export enum Role {
  CONSULTORTODOS = 'consultorTodos',
  CONSULTORCULTURA = 'consultorCultura',
  CONSULTORRECETA = 'consultorReceta',
  CONSULTORPRODUCTO = 'consultorProducto',
  CONSULTORCATEGORIA = 'consultorCategoria',
  CONSULTORPAIS = 'consultorPais',
  CONSULTORCIUDAD = 'consultorCiudad',
  CONSULTORRESTAURANTE = 'consultorRestaurante',
  CONSULTORESTRELLAMICHELIN = 'consultorEstrellaMichelin',
  ESCRITOR = 'escritor',
  ELIMINADOR = 'eliminador',
  ADMIN = 'admin',
}